requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0
Pillow==9.5.0
cairosvg==2.7.0  # For SVG conversion
numpy==1.24.3
pytube==15.0.0
pydub==0.25.1 